refactored types, refactored socket func to have rooms
