+++ backend/src/index.ts
@@ -3,6 +3,13 @@ import { config } from "dotenv";
 import express from "express";
 import { Server } from "socket.io";
 import { createServer } from "http";
+import {
+	RoomData,
+	SocketCursorData,
+	SocketDeleteData,
+	SocketDragTile,
+	UserData,
+} from "../types/socket.types";
 
 config();
 const app = express();
@@ -15,62 +22,6 @@ app.get("/", (req: express.Request, res: express.Response) =>
 	res.send("Hello from Server" + port),
 );
 
-
-
-export type UserData = {
-	roomId: string;
-	userName: string;
-	userId: string;
-	isHost: boolean;
-	cursorPos?: {
-		x: number;
-		y: number;
-	};
-};
-
-export type CursorData = {
-	x: number;
-	y: number;
-	remoteUser: string;
-};
-
-export type NewNode = {
-	id: string;
-	category: string;
-	src: string;
-	x: number;
-	y: number;
-};
-
-export type SocketDragTile = {
-	remoteUser: string;
-	tile: NewNode;
-	roomId: string;
-};
-
-export type SocketCursorData = {
-	x: number;
-	y: number;
-	remoteUser: string;
-	roomId: string;
-};
-
-//state object for each room
-export type TileData = {
-	tiles: {
-		id: string;
-		category: string;
-		src: string;
-		x: number;
-		y: number;
-	};
-};
-
-export type RoomData = {
-	roomId: string;
-	users: UserData[];
-	tiles?: TileData[];
-};
 const state: RoomData[] = [];
 
 app.get("/state", (req, res) => {
@@ -110,9 +61,13 @@ io.on("connection", (socket) => {
 	});
 
 	socket.on("tile-drop", (data: SocketDragTile) => {
-		console.log(data);
+		/**
+		 * check if the room exists
+		 * if it does, add the tile to the room
+		 * update the room state
+		 * emit the room data to the every user in the room
+		 */
 		const room = state.find((room) => room.roomId === data.roomId);
-		console.log(room);
 		if (room) {
 			room.tiles.push({
 				tiles: {
@@ -127,32 +82,66 @@ io.on("connection", (socket) => {
 		}
 	});
 
-	// state.cachedTiles.push(data.socketDragTile);
-	// socket.to(data.roomId).emit("tile-drop", data.socketDragTile);
-
 	// socket.on("tab-focus", (data: boolean) => {
 	// 	console.log("tab-focus", data);
 	// });
 
-	// socket.on("tile-drag", (data: SocketDragTileData) => {
-	// 	socket.broadcast.emit("tile-drag", data.socketDragTile);
-	// 	state.cachedTiles.forEach((cachedTile) => {
-	// 		if (cachedTile.tile.id === data.socketDragTile.tile.id) {
-	// 			cachedTile.tile.x = data.socketDragTile.tile.x;
-	// 			cachedTile.tile.y = data.socketDragTile.tile.y;
-	// 		}
-	// 	});
-	// });
-	// socket.on("cursor", (data: SocketCursorData) => {
-	// 	socket.to(data.roomId).emit("cursor", data.cursorPos);
-	// });
+	socket.on("tile-drag", (data: SocketDragTile) => {
+		/**
+		 * check if the room exists
+		 * if it does, check if the tile exists
+		 * if it does, update the tile position
+		 * update the room state
+		 * emit the room data to the every user in the room
+		 */
 
-	// socket.on("tile-delete", (data: string) => {
-	// 	socket.broadcast.emit("tile-delete", data);
-	// 	state.cachedTiles = state.cachedTiles.filter(
-	// 		(tile) => tile.tile.id !== data,
-	// 	);
-	// });
+		const room = state.find((room) => room.roomId === data.roomId);
+		if (room) {
+			const tile = room.tiles.find((tile) => tile.tiles.id === data.tile.id);
+			if (tile) {
+				tile.tiles.x = data.tile.x;
+				tile.tiles.y = data.tile.y;
+				io.to(data.roomId).emit("room-data", room);
+			}
+		}
+	});
+
+	socket.on("cursor", (data: SocketCursorData) => {
+		/**
+		 * check if the room exists
+		 * if it does, check if the user exists
+		 * if it does, update the user cursor position
+		 * update the room state
+		 * emit the room data to the every user in the room
+		 */
+		const room = state.find((room) => room.roomId === data.roomId);
+		if (room) {
+			const user = room.users.find((user) => user.userId === data.remoteUser);
+			if (user) {
+				user.cursorPos = {
+					x: data.x,
+					y: data.y,
+				};
+				io.to(data.roomId).emit("room-data", room);
+			}
+		}
+	});
+
+	socket.on("tile-delete", (data: SocketDeleteData) => {
+		/**
+		 * check if the room exists
+		 * if it does,then check if the tile exists
+		 * if it does, delete the tile from the room
+		 * update the room state
+		 * emit the room data to the every user in the room
+		 */
+
+		const room = state.find((room) => room.roomId === data.roomId);
+		if (room) {
+			room.tiles = room.tiles.filter((tile) => tile.tiles.id !== data.id);
+			io.to(data.roomId).emit("room-data", room);
+		}
+	});
 
 	socket.on("disconnect", () => {
 		/**
